style(feed page): rename colocated components, fit to style guide
